feat: Add wrapper initialization and proxy configuration
